pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prflow"
version = "0.1.0"
description = "CLI tool that automates PR preparation"
requires-python = ">=3.9"
dependencies = [
    "click>=8",
    "rich>=13",
    "PyYAML>=6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "pytest-mock>=3",
]

[project.scripts]
prflow = "prflow.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
